requirements.txt
asgiref==3.8.1
Django>=4.2.0
djangorestframework==3.15.0
sqlparse==0.5.3
tzdata==2025.2
coreapi==2.3.3
Markdown==3.5.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
PyJWT==2.8.0
